Fix: rename library exceptions folder
